Переключил на Blueprint схему
