+++ webapp/admin/__init__.py
@@ -4,9 +4,6 @@ from flask_login import current_user
 from flask import redirect, url_for
 
 
-
-
-
 class AdminView(ModelView):
     def is_accessible(self):
         return current_user.is_authenticated and current_user.is_admin
