+++ webapp/db.py
@@ -1,17 +1,6 @@
-from sqlalchemy import create_engine
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import scoped_session, sessionmaker
 from flask_sqlalchemy import SQLAlchemy
-from flask import Flask, current_app
-from webapp import config
 
-
-engine = create_engine(config['SQLALCHEMY_DATABASE_URI'])
-db_session = scoped_session(sessionmaker(bind=engine))
-
-Base = declarative_base()
-Base.query = db_session.query_property()
-
+db = SQLAlchemy()
