+++ webapp/templates/edit_profile.html
@@ -5,7 +5,7 @@
         <div class='col-4'>
         </div>
         <div class='col-8'>
-            <form action="{{ url_for('edit_adress') }}" method="POST">
+            <form action="{{ url_for('user.edit_adress') }}" method="POST">
                 {{ form.hidden_tag() }}
  
                 <div class='form-group'>
