+++ webapp/templates/login.html
@@ -2,7 +2,7 @@
 {% block content %}
 <div class='row'>
     <div class='col-4'>
-        <form action="{{ url_for('process_login') }}" method="POST">
+        <form action="{{ url_for('user.process_login') }}" method="POST">
             {{ form.hidden_tag() }}
             <div class='form-group'>
                 {{ form.email.label }}
