+++ webapp/templates/user_profile.html
@@ -14,7 +14,7 @@
             {% endif %}
         </div>
         <div class='col-4'>
-            <a href="{{ url_for('edit_profile', user_id=user.id) }}" class="btn btn-light btn-sm" role="button">Изменить адрес доставки</a>
+            <a href="{{ url_for('user.edit_profile', user_id=user.id) }}" class="btn btn-light btn-sm" role="button">Изменить адрес доставки</a>
         </div>
     </div>
 {% endblock %}
