+++ webapp/user/forms.py
@@ -1,7 +1,7 @@
 from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, SubmitField, BooleanField
 from wtforms.validators import DataRequired, Email, EqualTo, ValidationError, InputRequired
-from webapp.models import User
+from webapp.user.models import User, User_adress
 
 
 class LoginForm(FlaskForm):
@@ -20,7 +20,6 @@ class RegistrationForm(FlaskForm):
         users_count = User.query.filter_by(email=email.data).count()
         if users_count > 0:
             raise ValidationError('Пользователь с такой электронной почтой уже зарегистрирован')
-        
 
 class AddressForm(FlaskForm):
     city = StringField('Город', validators=[InputRequired()],  render_kw={"class": "form-control"})
